Remove some duplicate code in PTrace.
A lot of code is identical between the Linux, FreeBSD and Darwin
implementations of our ptrace abstraction. This commit moves a couple of
functions up into the POSIX implementation.

There are probably more methods that could be moved in the same way.
