+++ Headers/DebugServer2/Host/Darwin/PTrace.h
@@ -13,9 +13,6 @@
 
 #include "DebugServer2/Host/POSIX/PTrace.h"
 
-#include <sys/ptrace.h>
-#include <sys/types.h>
-
 namespace ds2 {
 namespace Host {
 namespace Darwin {
@@ -30,10 +27,6 @@ public:
 public:
   virtual ErrorCode traceMe(bool disableASLR);
   virtual ErrorCode traceThat(ProcessId pid);
-
-public:
-  virtual ErrorCode attach(ProcessId pid);
-  virtual ErrorCode detach(ProcessId pid);
 
 public:
   virtual ErrorCode kill(ProcessThreadId const &ptid, int signal);
@@ -73,15 +66,6 @@ public:
 protected:
   void initCPUState(ProcessId pid);
   void doneCPUState();
-
-protected:
-  template <typename CommandType, typename AddrType, typename DataType>
-  long wrapPtrace(CommandType request, pid_t pid, AddrType addr,
-                  DataType data) {
-    typedef int ptrace_request_t;
-    return ::ptrace(static_cast<ptrace_request_t>(request), pid,
-                    (char *)(uintptr_t)addr, (int)(uintptr_t)data);
-  }
 
 public:
   PTracePrivateData *_privateData;
